+++ lending/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.7.67",
-  genVersion: "2.499.0",
-  userAgent: "speakeasy-sdk/typescript 0.7.67 2.499.0 1.0.0 ryan-lending",
+  sdkVersion: "0.7.68",
+  genVersion: "2.500.5",
+  userAgent: "speakeasy-sdk/typescript 0.7.68 2.500.5 1.0.0 ryan-lending",
 } as const;